Closes #1859: Address review comments
