+++ app/src/system/java/org/mozilla/tv/firefox/ext/EngineView.kt
@@ -15,12 +15,13 @@ import android.webkit.WebView
 import mozilla.components.browser.engine.system.SystemEngineSession
 import mozilla.components.browser.session.SessionManager
 import mozilla.components.concept.engine.EngineView
-import org.mozilla.tv.firefox.ext.JS.CACHE_JS
-import org.mozilla.tv.firefox.ext.JS.CACHE_VAR
-import org.mozilla.tv.firefox.ext.JS.bodyElementFocused
-import org.mozilla.tv.firefox.ext.JS.noElementFocused
-import org.mozilla.tv.firefox.ext.JS.pauseVideo
-import org.mozilla.tv.firefox.ext.JS.sidebarFocused
+import org.mozilla.tv.firefox.ext.Js.BODY_ELEMENT_FOCUSED
+import org.mozilla.tv.firefox.ext.Js.CACHE_JS
+import org.mozilla.tv.firefox.ext.Js.CACHE_VAR
+import org.mozilla.tv.firefox.ext.Js.JS_OBSERVE_PLAYBACK_STATE
+import org.mozilla.tv.firefox.ext.Js.NO_ELEMENT_FOCUSED
+import org.mozilla.tv.firefox.ext.Js.PAUSE_VIDEO
+import org.mozilla.tv.firefox.ext.Js.SIDEBAR_FOCUSED
 import org.mozilla.tv.firefox.webrender.FocusedDOMElementCache
 import java.util.WeakHashMap
 
@@ -67,14 +68,14 @@ fun EngineView.evalJS(javascript: String, callback: ValueCallback<String>? = nul
 }
 
 fun EngineView.pauseAllVideoPlaybacks() {
-    evalJS(pauseVideo)
+    evalJS(PAUSE_VIDEO)
 }
 
-fun EngineView.domElementCache() {
+fun EngineView.cacheDomElement() {
     evalJS(CACHE_JS)
 }
 
-fun EngineView.domElementRestore() {
+fun EngineView.restoreDomElement() {
     evalJS("if ($CACHE_VAR) $CACHE_VAR.focus();")
 }
 
@@ -150,9 +151,9 @@ fun EngineView.seekTargetVideoToPosition(absolutePositionSeconds: Long) {
 fun EngineView.checkYoutubeBack(callback: ValueCallback<String>) {
     val shouldWeExitPage = """
                (function () {
-                    return $noElementFocused ||
-                        $bodyElementFocused ||
-                        $sidebarFocused;
+                    return $NO_ELEMENT_FOCUSED ||
+                        $BODY_ELEMENT_FOCUSED ||
+                        $SIDEBAR_FOCUSED;
                 })();
         """.trimIndent()
 
@@ -263,138 +264,3 @@ private class EngineViewExtension(private val engineView: EngineView) {
                 null
             }
 }
-
-/**
- * This script will:
- * - Add playback state change listeners to all <video>s in the DOM; it uses a mutation
- *   observer to attach listeners to new <video> nodes as well
- * - On playback state change, notify Java about the current playback state
- * - Prevent this script from being injected more than once per page
- *
- * Note that `//` style comments are not supported in `evalJS`.
- *
- * Development tips:
- * - This script was written using Typescript with Visual Studio Code: it may be easier to modify
- *   it by copy-pasting it back-and-forth.
- * - Iterating on the Fire TV is slow: you can speed it up by making this a WebExtension content
- *   script and testing on desktop
- * - For a list of HTMLMediaElement (i.e. video) events, like 'ratechange', see the w3c's HTML5 video
- *   page: https://www.w3.org/2010/05/video/mediaevents.html
- */
-private val JS_OBSERVE_PLAYBACK_STATE = """
-var _firefoxTV_playbackStateObserverJava;
-var _firefoxTV_isPlaybackStateObserverLoaded;
-(function () {
-    /* seeking will send "pause, play" and so is covered here. */
-    const PLAYBACK_STATE_CHANGE_EVENTS = ['play', 'pause', 'ratechange'];
-    const MILLIS_BETWEEN_PLAYBACK_STATE_SYNC_BY_TIME = 1000 * 10 /* seconds */;
-
-    const javaInterface = _firefoxTV_playbackStateObserverJava;
-    if (!javaInterface) {
-        console.error('Cannot sync playback state to Java: JavascriptInterface is not found.');
-    }
-
-    const videosWithListeners = new Set();
-
-    let playbackStateSyncIntervalID;
-
-    function onDOMChangedForVideos() {
-        addPlaybackStateListeners();
-        syncPlaybackState();
-    }
-
-    function addPlaybackStateListeners() {
-        document.querySelectorAll('video').forEach(videoElement => {
-            if (videosWithListeners.has(videoElement)) { return; }
-            videosWithListeners.add(videoElement);
-
-            PLAYBACK_STATE_CHANGE_EVENTS.forEach(event => {
-                videoElement.addEventListener(event, syncPlaybackState);
-            });
-        });
-    }
-
-    function syncPlaybackState() {
-        let isVideoPresent;
-        let isPlaying;
-        let positionSeconds;
-        let playbackRate; /* 0.5, 1, etc. */
-        const maybeTargetVideo = getPlayingVideoOrFirstInDOMOrNull();
-        if (maybeTargetVideo) {
-            isVideoPresent = true;
-            isPlaying = !maybeTargetVideo.paused;
-            positionSeconds = maybeTargetVideo.currentTime;
-            playbackRate = maybeTargetVideo.playbackRate;
-        } else {
-            isVideoPresent = false;
-            isPlaying = false;
-            positionSeconds = null;
-            playbackRate = null;
-        }
-
-        schedulePlaybackStateSyncInterval(isPlaying);
-
-        javaInterface.syncPlaybackState(isVideoPresent, isPlaying, positionSeconds, playbackRate);
-    }
-
-    /**
-     * When a video is playing, schedules a function to repeatedly sync the playback state;
-     * cancels it when there is no video playing.
-     *
-     * Java and JavaScript increment the current playback position independently and run the risk of
-     * getting out of sync (e.g. upon buffering). We could try to handle the buffering case specifically
-     * but its state is difficult to identify with and syncing periodically is a better general solution.
-     * We don't sync with the video's 'timeupdate' event because it's called very frequently and could
-     * detract from performance.
-     */
-    function schedulePlaybackStateSyncInterval(isVideoPlaying) {
-        if (isVideoPlaying && !playbackStateSyncIntervalID) {
-            playbackStateSyncIntervalID = setInterval(syncPlaybackState,
-                MILLIS_BETWEEN_PLAYBACK_STATE_SYNC_BY_TIME);
-
-        } else if (!isVideoPlaying && playbackStateSyncIntervalID) {
-            clearInterval(playbackStateSyncIntervalID);
-            playbackStateSyncIntervalID = null;
-        }
-    }
-
-    function getPlayingVideoOrFirstInDOMOrNull() {
-        const maybePlayingVideo = Array.from(document.querySelectorAll('video')).find(video => !video.paused);
-        if (maybePlayingVideo) { return maybePlayingVideo; }
-
-        /* If there are no playing videos, just return the first one. */
-        return document.querySelector('video');
-    }
-
-    function nodeContainsVideo(node) {
-        return node.nodeName.toLowerCase() === 'video' ||
-                ((node instanceof Element) && !!node.querySelector('video'));
-    }
-
-    const documentChangedObserver = new MutationObserver(mutationList => {
-        const wasVideoAdded = mutationList.some(mutation => {
-            return mutation.type === 'childList' &&
-                    (Array.from(mutation.addedNodes).some(nodeContainsVideo) ||
-                            Array.from(mutation.removedNodes).some(nodeContainsVideo));
-        });
-
-        if (wasVideoAdded) {
-            /* This may traverse the whole DOM so let's only call it if it's necessary. */
-            onDOMChangedForVideos();
-        }
-    });
-
-    /* Sometimes the script is evaluated more than once per page:
-     * only inject code with side effects once. */
-    if (!_firefoxTV_isPlaybackStateObserverLoaded) {
-        _firefoxTV_isPlaybackStateObserverLoaded = true;
-
-        documentChangedObserver.observe(document, {subtree: true, childList: true});
-
-        /* The DOM is changed from blank to filled for the initial page load.
-         * While the function name assumes videos are present, checking for
-         * videos is as expensive as calling the function so we just call it. */
-        onDOMChangedForVideos();
-    }
-})();
-""".trimIndent()
